Bitvmx script generation entities creation
