+++ bitvmx_protocol_library/script_generation/entities/bitcoin_script.py
