+++ bitvmx_protocol_library/script_generation/entities/bitcoin_script_list.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Union
 
 from bitcoinutils.keys import PublicKey
 
-from scripts.bitcoin_script import BitcoinScript
+from bitvmx_protocol_library.script_generation.entities.bitcoin_script import BitcoinScript
 
 
 class BitcoinScriptList:
