+++ bitvmx_protocol_library/script_generation/services/commit_search_hashes_script_generator_service.py
@@ -1,18 +1,24 @@
+from typing import List, Optional
+
 from bitcoinutils.keys import PublicKey
 
-from scripts.bitcoin_script import BitcoinScript
+from bitvmx_protocol_library.script_generation.entities.bitcoin_script import BitcoinScript
 from winternitz_keys_handling.scripts.confirm_single_word_script_generator_service import (
     ConfirmSingleWordScriptGeneratorService,
 )
 from winternitz_keys_handling.scripts.verify_digit_signature_nibbles_service import (
     VerifyDigitSignatureNibblesService,
 )
+from winternitz_keys_handling.scripts.verify_digit_signature_single_word_service import (
+    VerifyDigitSignatureSingleWordService,
+)
 
 
-class ExecutionTraceScriptGeneratorService:
+class CommitSearchHashesScriptGeneratorService:
 
     def __init__(self):
         self.verify_input_nibble_message_from_public_keys = VerifyDigitSignatureNibblesService()
+        self.verify_input_single_word_from_public_keys = VerifyDigitSignatureSingleWordService()
         self.confirm_single_word_script_generator_service = (
             ConfirmSingleWordScriptGeneratorService()
         )
@@ -21,11 +27,11 @@ class ExecutionTraceScriptGeneratorService:
         self,
         signature_public_keys,
         public_keys,
-        trace_words_lengths,
+        n0,
         bits_per_digit_checksum,
-        amount_of_bits_choice,
-        prover_choice_public_keys,
-        verifier_choice_public_keys,
+        amount_of_bits_choice: Optional[int] = None,
+        prover_choice_public_keys: Optional[List[str]] = None,
+        verifier_choice_public_keys: Optional[List[str]] = None,
     ):
         script = BitcoinScript()
 
@@ -33,17 +39,23 @@ class ExecutionTraceScriptGeneratorService:
             self.verify_input_nibble_message_from_public_keys(
                 script,
                 public_keys[i],
-                trace_words_lengths[i],
+                n0,
                 bits_per_digit_checksum,
                 to_alt_stack=True,
             )
+        if (
+            amount_of_bits_choice is not None
+            and prover_choice_public_keys is not None
+            and verifier_choice_public_keys is not None
+        ):
+            self.confirm_single_word_script_generator_service(
+                script,
+                amount_of_bits_choice,
+                prover_choice_public_keys,
+                verifier_choice_public_keys,
+            )
 
-        self.confirm_single_word_script_generator_service(
-            script,
-            amount_of_bits_choice,
-            prover_choice_public_keys,
-            verifier_choice_public_keys,
-        )
+        # script.append("OP_CODESEPARATOR")
 
         for signature_public_key in reversed(signature_public_keys):
             script.extend(
