+++ bitvmx_protocol_library/script_generation/services/hash_result_script_generator_service.py
@@ -1,6 +1,6 @@
 from bitcoinutils.keys import PublicKey
 
-from scripts.bitcoin_script import BitcoinScript
+from bitvmx_protocol_library.script_generation.entities.bitcoin_script import BitcoinScript
 from winternitz_keys_handling.scripts.verify_digit_signature_nibbles_service import (
     VerifyDigitSignatureNibblesService,
 )
