+++ bitvmx_protocol_library/script_generation/services/trigger_generic_challenge_script_generator_service.py
@@ -1,6 +1,6 @@
 from bitcoinutils.keys import PublicKey
 
-from scripts.bitcoin_script import BitcoinScript
+from bitvmx_protocol_library.script_generation.entities.bitcoin_script import BitcoinScript
 from winternitz_keys_handling.scripts.confirm_nibbles_script_generator_service import (
     ConfirmNibblesScriptGeneratorService,
 )
