+++ prover_app/main.py
@@ -26,7 +26,7 @@ from blockchain_query_services.services.blockchain_query_services_dependency_inj
 )
 from blockchain_query_services.services.mutinynet_api.faucet_service import FaucetService
 from prover_app.config import protocol_properties
-from scripts.scripts_dict_generator_service import ScriptsDictGeneratorService
+from bitvmx_protocol_library.script_generation.scripts_dict_generator_service import ScriptsDictGeneratorService
 from transactions.enums import TransactionProverStepType
 from transactions.generate_signatures_service import GenerateSignaturesService
 from transactions.publication_services.execution_challenge_transaction_service import (
