+++ transactions/publication_services/execution_challenge_transaction_service.py
@@ -9,7 +9,7 @@ from blockchain_query_services.services.blockchain_query_services_dependency_inj
     broadcast_transaction_service,
     transaction_info_service,
 )
-from scripts.services.execution_challenge_script_list_generator_service import (
+from bitvmx_protocol_library.script_generation.services.execution_challenge_script_list_generator_service import (
     ExecutionChallengeScriptListGeneratorService,
 )
