+++ transactions/publication_services/publish_choice_search_transaction_service.py
@@ -7,7 +7,7 @@ from blockchain_query_services.services.blockchain_query_services_dependency_inj
     broadcast_transaction_service,
     transaction_info_service,
 )
-from scripts.services.commit_search_choice_script_generator_service import (
+from bitvmx_protocol_library.script_generation.services.commit_search_choice_script_generator_service import (
     CommitSearchChoiceScriptGeneratorService,
 )
 from winternitz_keys_handling.services.generate_witness_from_input_single_word_service import (
