+++ transactions/publication_services/publish_hash_transaction_service.py
@@ -11,7 +11,7 @@ from bitvmx_execution.services.execution_trace_query_service import ExecutionTra
 from blockchain_query_services.services.blockchain_query_services_dependency_injection import (
     broadcast_transaction_service,
 )
-from scripts.services.hash_result_script_generator_service import HashResultScriptGeneratorService
+from bitvmx_protocol_library.script_generation.services.hash_result_script_generator_service import HashResultScriptGeneratorService
 from winternitz_keys_handling.services.generate_witness_from_input_nibbles_service import (
     GenerateWitnessFromInputNibblesService,
 )
