+++ transactions/publication_services/publish_trace_transaction_service.py
@@ -7,7 +7,7 @@ from blockchain_query_services.services.blockchain_query_services_dependency_inj
     broadcast_transaction_service,
     transaction_info_service,
 )
-from scripts.services.execution_trace_script_generator_service import (
+from bitvmx_protocol_library.script_generation.services.execution_trace_script_generator_service import (
     ExecutionTraceScriptGeneratorService,
 )
 from winternitz_keys_handling.services.generate_witness_from_input_nibbles_service import (
