+++ transactions/publication_services/trigger_execution_challenge_transaction_service.py
@@ -5,7 +5,7 @@ from bitcoinutils.utils import ControlBlock
 from blockchain_query_services.services.blockchain_query_services_dependency_injection import (
     broadcast_transaction_service,
 )
-from scripts.services.trigger_generic_challenge_script_generator_service import (
+from bitvmx_protocol_library.script_generation.services.trigger_generic_challenge_script_generator_service import (
     TriggerGenericChallengeScriptGeneratorService,
 )
 from winternitz_keys_handling.services.generate_witness_from_input_nibbles_service import (
