+++ transactions/publication_services/trigger_protocol_transaction_service.py
@@ -9,7 +9,7 @@ from bitvmx_execution.services.execution_trace_query_service import ExecutionTra
 from blockchain_query_services.services.blockchain_query_services_dependency_injection import (
     broadcast_transaction_service,
 )
-from scripts.services.trigger_protocol_script_generator_service import (
+from bitvmx_protocol_library.script_generation.services.trigger_protocol_script_generator_service import (
     TriggerProtocolScriptGeneratorService,
 )
