+++ transactions/publication_services/trigger_wrong_hash_challenge_transaction_service.py
@@ -1,4 +1,4 @@
-from scripts.services.trigger_generic_challenge_script_generator_service import (
+from bitvmx_protocol_library.script_generation.services.trigger_generic_challenge_script_generator_service import (
     TriggerGenericChallengeScriptGeneratorService,
 )
 from winternitz_keys_handling.services.generate_witness_from_input_nibbles_service import (
