+++ transactions/transaction_generator_from_public_keys_service.py
@@ -3,7 +3,7 @@ from bitcoinutils.transactions import Transaction, TxInput, TxOutput
 
 from bitvmx_protocol_library.config import common_protocol_properties
 from bitvmx_protocol_library.enums import BitcoinNetwork
-from scripts.scripts_dict_generator_service import ScriptsDictGeneratorService
+from bitvmx_protocol_library.script_generation.scripts_dict_generator_service import ScriptsDictGeneratorService
 
 
 class TransactionGeneratorFromPublicKeysService:
