+++ verifier_app/main.py
@@ -17,7 +17,7 @@ from blockchain_query_services.services.blockchain_query_services_dependency_inj
     transaction_info_service,
 )
 from prover_app.config import protocol_properties
-from scripts.scripts_dict_generator_service import ScriptsDictGeneratorService
+from bitvmx_protocol_library.script_generation.scripts_dict_generator_service import ScriptsDictGeneratorService
 from transactions.enums import TransactionVerifierStepType
 from transactions.generate_signatures_service import GenerateSignaturesService
 from transactions.publication_services.publish_choice_search_transaction_service import (
